Move classes related the authentication into auth package
