+++ gnd/src/main/java/com/google/android/gnd/system/auth/AuthenticationModule.java
@@ -1,6 +1,8 @@
-package com.google.android.gnd.system;
+package com.google.android.gnd.system.auth;
 
 import com.google.android.gnd.inject.ActivityScoped;
+import com.google.android.gnd.system.auth.AuthenticationManager;
+import com.google.android.gnd.system.auth.AuthenticationManagerImpl;
 import dagger.Binds;
 import dagger.Module;
