+++ gnd/src/main/java/com/google/android/gnd/system/auth/SignInState.java
@@ -1,4 +1,4 @@
-package com.google.android.gnd.system;
+package com.google.android.gnd.system.auth;
 
 import com.google.android.gnd.model.User;
 import com.google.android.gnd.rx.ValueOrError;
